backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI
openai==1.3.7

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Python環境
python-dotenv==1.0.0
